refactored ExpenseList component
